# TMDB_CLONE/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XpoGo Streaming API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface TmdbItem {
  id: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  name?: string | null;
  overview: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  vote_count: number;
  /** @nullable */
  release_date?: string | null;
  /** @nullable */
  first_air_date?: string | null;
  /** @nullable */
  media_type?: string | null;
  genre_ids?: number[];
  popularity: number;
}

export interface TmdbListResponse {
  page: number;
  results: TmdbItem[];
  total_pages: number;
  total_results: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface GenreListResponse {
  genres: Genre[];
}

export type TmdbMovieDetailVideosResultsItem = { [key: string]: unknown };

export type TmdbMovieDetailVideos = {
  results?: TmdbMovieDetailVideosResultsItem[];
};

export type TmdbMovieDetailCreditsCastItem = { [key: string]: unknown };

export type TmdbMovieDetailCreditsCrewItem = { [key: string]: unknown };

export type TmdbMovieDetailCredits = {
  cast?: TmdbMovieDetailCreditsCastItem[];
  crew?: TmdbMovieDetailCreditsCrewItem[];
};

export interface TmdbMovieDetail {
  id: number;
  title: string;
  overview: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  vote_count: number;
  release_date: string;
  /** @nullable */
  runtime?: number | null;
  genres: Genre[];
  status: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  homepage?: string | null;
  /** @nullable */
  imdb_id?: string | null;
  videos?: TmdbMovieDetailVideos;
  credits?: TmdbMovieDetailCredits;
  similar?: TmdbListResponse;
  recommendations?: TmdbListResponse;
}

export interface TmdbEpisode {
  id: number;
  name: string;
  overview: string;
  episode_number: number;
  season_number: number;
  /** @nullable */
  still_path?: string | null;
  /** @nullable */
  air_date?: string | null;
  vote_average: number;
  /** @nullable */
  runtime?: number | null;
}

export interface TmdbSeasonDetail {
  id: number;
  name: string;
  overview: string;
  season_number: number;
  episode_count: number;
  /** @nullable */
  air_date?: string | null;
  /** @nullable */
  poster_path?: string | null;
  episodes: TmdbEpisode[];
}

export type TmdbTvDetailSeasonsItem = { [key: string]: unknown };

export type TmdbTvDetailVideosResultsItem = { [key: string]: unknown };

export type TmdbTvDetailVideos = {
  results?: TmdbTvDetailVideosResultsItem[];
};

export type TmdbTvDetailCreditsCastItem = { [key: string]: unknown };

export type TmdbTvDetailCreditsCrewItem = { [key: string]: unknown };

export type TmdbTvDetailCredits = {
  cast?: TmdbTvDetailCreditsCastItem[];
  crew?: TmdbTvDetailCreditsCrewItem[];
};

export interface TmdbTvDetail {
  id: number;
  name: string;
  overview: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  vote_count: number;
  first_air_date: string;
  /** @nullable */
  last_air_date?: string | null;
  number_of_seasons: number;
  number_of_episodes: number;
  genres: Genre[];
  seasons?: TmdbTvDetailSeasonsItem[];
  status: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  homepage?: string | null;
  videos?: TmdbTvDetailVideos;
  credits?: TmdbTvDetailCredits;
  similar?: TmdbListResponse;
  recommendations?: TmdbListResponse;
}

export interface PlayerSettings {
  server: string;
  domainAd: string;
  color: string;
  autoplay: boolean;
  back: boolean;
}

export interface UiSettings {
  posterSize: string;
  fontSize: string;
  accentColor: string;
  fontColor: string;
}

export interface AppSettings {
  player: PlayerSettings;
  ui: UiSettings;
}

export interface Ad {
  id: string;
  title: string;
  url: string;
  imageUrl: string;
  active: boolean;
  position: string;
}

export interface AdBody {
  title: string;
  url: string;
  imageUrl: string;
  active: boolean;
  position: string;
}

export interface CustomMovie {
  id: string;
  tmdbId: number;
  type: string;
  title: string;
  posterPath: string;
  featured: boolean;
  category: string;
}

export interface CustomMovieBody {
  tmdbId: number;
  type: string;
  title: string;
  posterPath: string;
  featured: boolean;
  category: string;
}

export interface EmbedEpisode {
  episode: number;
  url: string;
  sub: string;
}

export interface EmbedSeason {
  season: number;
  episodes: EmbedEpisode[];
}

export interface EmbedEntry {
  id: string;
  tmdbId: number;
  /** movie or series */
  type: string;
  title: string;
  /** @nullable */
  posterPath?: string | null;
  /** Subtitle language e.g. Sub Indo, Sub English, Dub */
  sub: string;
  /**
   * Full embed URL for movie
   * @nullable
   */
  url?: string | null;
  /** For series: list of seasons with episodes */
  seasons?: EmbedSeason[];
  active: boolean;
  createdAt: string;
}

export interface EmbedEntryBody {
  tmdbId: number;
  /** movie or series */
  type: string;
  title: string;
  /** @nullable */
  posterPath?: string | null;
  /** Subtitle language e.g. Sub Indo, Sub English, Dub */
  sub: string;
  /**
   * Full embed URL for movie
   * @nullable
   */
  url?: string | null;
  /** For series: list of seasons with episodes */
  seasons?: EmbedSeason[];
  active: boolean;
}

export type GetTrendingParams = {
  media_type?: string;
  time_window?: string;
  page?: number;
};

export type GetPopularMoviesParams = {
  page?: number;
};

export type GetPopularTvParams = {
  page?: number;
};

export type GetTopRatedMoviesParams = {
  page?: number;
};

export type GetTopRatedTvParams = {
  page?: number;
};

export type SearchTmdbParams = {
  query: string;
  page?: number;
};

export type GetMoviesByGenreParams = {
  page?: number;
};