# TMDB_CLONE/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XpoGo Streaming API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ads:
    get:
      operationId: getAds
      tags: [ads]
      summary: Get all ads
      responses:
        "200":
          description: Ads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [ads]
      summary: Create ad (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdBody"
      responses:
        "201":
          description: Created ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /ads/{id}:
    put:
      operationId: updateAd
      tags: [ads]
      summary: Update ad (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdBody"
      responses:
        "200":
          description: Updated ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    delete:
      operationId: deleteAd
      tags: [ads]
      summary: Delete ad (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /custom-movies:
    get:
      operationId: getCustomMovies
      tags: [movies]
      summary: Get custom movie list from Firebase
      responses:
        "200":
          description: Custom movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomMovie"
    post:
      operationId: createCustomMovie
      tags: [movies]
      summary: Add movie to custom list (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomMovieBody"
      responses:
        "201":
          description: Added movie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomMovie"
  /custom-movies/{id}:
    delete:
      operationId: deleteCustomMovie
      tags: [movies]
      summary: Remove from custom list (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
components:
  schemas:
...
    Ad:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
        imageUrl:
          type: string
        active:
          type: boolean
        position:
          type: string
      required:
        - id
        - title
        - url
        - imageUrl
        - active
        - position
    AdBody:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        imageUrl:
          type: string
        active:
          type: boolean
        position:
          type: string
      required:
        - title
        - url
        - imageUrl
        - active
        - position
    CustomMovie:
      type: object
      properties:
        id:
          type: string
        tmdbId:
          type: integer
        type:
          type: string
        title:
          type: string
        posterPath:
          type: string
        featured:
          type: boolean
        category:
          type: string
      required:
        - id
        - tmdbId
        - type
        - title
        - posterPath
        - featured
        - category
    CustomMovieBody:
      type: object
      properties:
        tmdbId:
          type: integer
        type:
          type: string
        title:
          type: string
        posterPath:
          type: string
        featured:
          type: boolean
        category:
          type: string
      required:
        - tmdbId
        - type
        - title
        - posterPath
        - featured
        - category
...